[Hunter] Move Kill Command to BM
